pyproject.toml
[tool.poetry]
name = "ag-hr-automation"
version = "0.1.0"
description = "This is the backend for HR-automation for Agilisys."
authors = ["Deepak, Anisha, Maitree"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
python-decouple = "^3.8"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
sqlalchemy = "2.0.32"
psycopg2-binary = "^2.9.10"
requests = "2.25.1"
jsonb = "^1.0.0"
botframework-connector = "4.14.0"
botbuilder-schema = "4.14.0"
botbuilder-core = "4.14.0"
aiohttp = "^3.8.1"
botbuilder-dialogs = "4.14.0"
emoji = "0.6.0"
recognizers-text = "^1.0.2a2"
#openai = "1.54.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
